Rename set_cursor to position and clean up
The set_cursor method was renamed to position to be consistent
with the HD44780 library.  The font is now declared as a const
and each byte has been reversed to avoid doing the reversal at
runtime.  The write_data method is now declared with const data
input as it should have been from the beginning.